Fixup initramfs rebuild process
At the time the DKMS kernel module is built and installed,
it's entirely possible that the utilities expected to be included
are not yet installed. Thus, DKMS will no longer produce the initramfs
and it'll be done by dattobd-utils instead. This way, dbdctl and friends
will be installed and properly included in the initramfs.