Use OrderWithRequires to ensure kernel is first in same transaction
With RPM 4.9.0 and newer, it's possible to give transaction hints to enforce
some kind of ordering for transactions using the OrderWithRequires statement.
More info: http://rpm.org/wiki/Releases/4.9.0#package-building

We can use this to ensure that if kernel-devel/kernel-syms is in the same transaction
as the DKMS module upgrade, it will be installed first before processing the kernel
module, ensuring that DKMS will be able to successfully build against the new kernel
being installed.