Fix falloc_size for info ioctl
If the size of the cow file was smaller than 1 MiB, sd_falloc_size would
(correctly) report the size as 0. This value was then converted to bytes
and passed to the dattobd_info struct for the info ioctl, meaning it
would incorrectly be set to zero. Use sd_cow->file_max instead, which
already holds the value in bytes.