Install the headers in packaging and fix Makefile references to headers
In order to make the devel subpackage useful, we need to include the header
files that allow applications to use functions defined in libdattobd.

This necessitated adjusting the files so that they are searched for properly.

In addition, the Makefile is now adjusted so that the search paths are correct
to be able to build the whole thing.