Add support for running the feature tests against non-running kernel
In commit 1fd030c09f1e6ef2f92f428f7e0e609ad66bf18b, we changed the
packaging such that the DKMS common scriptlet is used to build and
install the module, which enabled DKMS to be correctly invoked in
cases like building Linux distribution images or doing kernel upgrades.

However, the kernel feature test scripts used to configure the build of
the kernel module continued to assume that the running kernel is always
the correct target, which is not only wrong, but can lead to false failures
for supported kernels as the build is configured for a kernel with a different
feature set and ABI than what the target kernel is.

With this change, the feature test scripts are now guided to the correct
kernel module build trees that was passed in by DKMS or by the user,
which should fix the remaining major issues with having the module build
properly on kernel or distribution upgrades.