umount: Don't fail if device is read-only
During shutdown, many distros will now have a read-only mount of the
root file system somewhere else. The post-unmount check was overly
strict in this case, as it considered being able to get *any* reference
to the device as an error. Specify the check to only be an error if the
device is read-write.

Resolves #60