Track the number of changed blocks
Add a field that can be used to calculate how much data has changed
since the last snapshot. This information is exposed in both the proc
file and through the info ioctl.

For compatibility, a version is also added to the header. Devices with
unversioned headers will continue to work, without tracking changes.
The driver will automatically upgrade to the newer format on the next
snapshot.