init: Make root device read-only during mount
File systems such as ext4 may mount a volume read-write if the log is
dirty, even is `-o ro` is specified. Per the mount man page, set the
block device read-only before mount and read-write after unmount.

Rewrite the initramfs-tools and initrd scripts to be consistent with the
dracut version.