INSTALL.md: Change Debian/Ubuntu install steps to use public key file
We've encountered a number of issues with fetching GPG keys from
key servers. As currently supported versions of Debian and Ubuntu
have recent enough versions of APT to be able to fetch and load
the key from a remote file, we're switching to that.

This also sort of aligns us with how the GPG key is loaded on
RPM based distributions.