Move custom bio flag from bi_flags to bi_opf
With the follow commits upstream

- torvalds/linux@c0acf12a50c2b ("block: shrink bio size again")
- torvalds/linux@399254aaf4892 ("block: add BIO_NO_PAGE_REF flag")

there is no more space in bi_flags. We were using a bit in there to mark
bios as already tracked so that we would not continuously COW them upon
resubmitting.

Steal a bit from bi_rw/bi_opf for this purpose instead, which still has
a few to spare.