dist: Add gcc and make dependencies to ensure package build works
On recent Fedora releases, we are observing that tools like make
are not necessarily always installed on the system when
DKMS is installed. Since we need make for the Makefiles in dattobd,
we will now directly require it.

Moreover, we now specify gcc as build dependency because it is no
longer part of the minimal build environment in Fedora.