Add info to INSTALL.md about kernel-devel dependency (#56)
It would be great if it could be possible to add dependency of the 
kernel-devel-CURRENT_KERNEL_VER to dkms-elastio-snap package.
Unfortunately there is no solution how to add a variable to the 'Requires: package = version_var'
and detect it during the installation. The version have to be a constant.
Hence just updating INSTALL.md with instantiation to install `kernel-devel-$(uname -r) ` if running
not a fresh kernel.