Linux kernel 5.8 support (#72)
* Linux kernel 5.8 support

A block device requests queue has a function 'make_request_fn' which is
used in the driver to send requests to the device underlying to a
snapshot. Suddenly, starting from the kernel 5.8, this function has
become empty (NULL) in most cases.
Used call of the similar function 'blk_mq_make_request' in case if the
'make_request_fn' is not set. This call is wrapped into the
elastio_snap_null_mrf function with the memory barrier before call
blk_mq_make_request. Otherwise driver hangs on the snapshot destroy
operation.

Resolves #44

Note: build of the driver on the Linux kernel 5.8 has been already fixed
in the #48, commit 96d63d.

* Added possibility to use real block device in tests

Now name of the device can be specified to the tests/elio-test.sh
script as an argument.
This is useful in case of some issues related to the loop device.
Actually it's a workaround for the #71

* Enabled builds and tests for Fedora 31 with the kernel 5.8

Here is a workaround for the issue #71 has been used.
Another disk has been attached to the Fedora 31 VM especially for the
tests which are failing with the loop device (used by default).