Fix ability to take snapshots from the multiple devices at one time on kernels 5.9+ (#112)
This change helps to avoid hack with the replacement of the `const` member
function `submit_bio` in the `block_device_operations` structure.
So, now there is no need to disable/enable page protection when
tracing is started/stopped. Entire struct for the traced device is replaced now.
As experiment showed, this struct can be shared between multiple
devices. That was a root cause of the bug.

Also introduced new func `find_orig_fops` which is very like the
existing func `find_orig_mrf`. This is done to reduce count of
precompilation directives in the original function, which became
unreadable at all if still combined for kernels older and newer than 5.9.

Resolves #100