Add support for non-resident cow files
Cow files can now exist off of the disk it is cow-ing. This comes with a
few limitations:
1. You cannot unmount a device with an active cow file on it. You would
   have to destroy or unmount the device being snapshotted by that cow
   file. DattoBD will hold the file handle to the cow file and will
   prevent the unmount.
3. Don't create loops with the snapshotted devices. If device A's cow
   file exists on device B, and device B's cow file is on device A, then
   you wouldn't be able to unmount either device or shutdown the
   machine. You'd have to destroy either snapshot to be able to unmount.

This commit also adds some fields to the /proc/datto-info device. The
`resident` field is a 0/1 false/true value that answers the question of
whether the cow file is on the device it is tracking. The
`full_cow_path` field represents the full path to the cow device, rather
than `cow_file`'s relative-to-mountpoint path. You are encouraged to use
the new `full_cow_path` field; `cow_file` will be kept, but deprecated,
for backwards compatibility.