Support LVM and software RAID  (#144)
The main idea of the fix to support  LVM and software RAID is to use `bdops`instead of `queue`
for identification of the proper device in the `tracing_mrf`, `find_orig_mrf` etc.

Also implemented `bio_free_pages` compat to use the original function in the `bio_free_clone`
to avoid kernel panic in case if CoW file is running out space.

Implemented possibility to run existing tests on the mirrored LVM and raid devices. 
`elio-test.sh` now accepts arguments `--lvm` and `--raid`.
The tests for the mirrored device require 2 devices to build a mirror. So, now tests are using
2 loopback devices under the hood, when executed with the `-l` or `-r` parameters.
 
The tests are added to CI. However there are some issues discovered on on LVM with XFS,
so some tests for LVM and RAID are disabled for XFS with the references to the appropriate
issues.

Closes #89 
Closes #119