Fixed attempt to access beyond end of device (#161)
This fix closes #150 and properly approaches the integration of the v5.16 Linux kernel.

Previously, during the integration of the v5.16 kernel, we had an issue with the partition
mounting (test_modify_origin couldn't pass successfully). This issue was caused by the
wrong sector offset calculation. Then it was [incorrectly] considered to be a new, changed
approach to the sector calculation in the modern kernel. Now we see that understanding
wasn't correct.

During debugging of issue #150, a similar behavior's been found on v5.15 (ubuntu 22.04,
arm64). It was found that the issue is not related to the 5.16. Hence, the solution made 
during that time was rather a workaround and didn't actually explain the nature of the 
problem.

After proper debugging, it was seen that some bio requests were processed OK, while 
the others had the same offset problem. It was suspected that the bug may be related 
to the bio's nature (depending on the issuing entity). This was confirmed: indeed, the 
problem was with the absent flags, that have been introduced after v5.11 (or somewhere 
around). This explains random failures: it seems that sometimes bio requests were issued 
with proper partition remapping, and sometimes not. These random differences made up 
the problem with the wrong offsets.

This was fixed by copying the relevant flags in make_read_clone(). Additional compats for 
new kernels have been added as well.

Closes #150