Removed code related to the bdev ownership change functionality (#165)
Previously, in issue #134 we added the code dedicated to gaining and
restoring the ownership over the block device. This has been implemented
to make sure that the right entity (elastio_snap or parent driver) owns
the device. As we have changed the ownership logic in the fix for #149,
we can now revert this functionality and automatically fix this issue.

Resolves #157