Driver fails to unload on Ubuntu 22.04 (#172)
It was seen in #169 that in rare cases driver fails to unload on Ubuntu 22.04. This happens because, due to some reason unknown, the kernel increments the module refcount via the owner field. The issue was reproduced with CI tests only, there was no scenario found to reproduce it locally to debug deeper with additional tools.

In light of this, it was decided to comment the owner field, which doesn't seem to be causing additional stability issues on any distributive. Apart from that, additional retries were added to handle the module unloading process.

The fix has been deeply tested and proven itself as stable on all distros and architectures.

Closes #169