Implement functionality to switch on/of EIO propagation on snap read (#187)
-= Implement functionality to switch on/of EIO propagation on snap read =-

Added an optional `-i` argument to `elioctl` to switch off IO errors
propagation on snapshot read when the snap is in the failed state.
The same arg has been added to the library functions to setup a snap and
transition to a snap.
As result, the entire error handling has been refactored a bit.
A've got rid of `sd_memory_fail_code` and `sd_cow_fail_state` snapshot
struct members. The error managing with usage of three error states, two
mentioned plus `sd_fail_code` was superfluous. Moreover, in case of
the snap dev memory mapping, IO errors were not propagated just in case
of the 2 errors: mem is too low and COW file size exceeded. Now all
other possible errors should not kill userspace app with `SIGBUS` if it
reads the snap as a memory-mapped file.

The `ignore_snap_error` parameter was added to the setup and reload
CLI commands, library methods and structures for IOCTLs.
Also `ignore_snap_errors` is shown in `/proc/elastio-snap-info` and
in the output of the info IOCTL.

-= Other changes =-
- Fixed compilation warning on some kernel versions.
  `atomic64_read` returns `long int` but not `long long int` on the
  kernel 4.9 and older for amd64 and even on some newer kernels for arm64.
  See https://lore.kernel.org/lkml/CAK8P3a1PpgjRKFemEi7mLuECh1srs8bbzoc07RSrFsx+febjaQ@mail.gmail.com/T/

- Improved an error check on the low memory detection
- Add tests for reload commands by @kgermanov 
  The test `test_reload_verified_snapshot` is disabled on 3rd kernels of Debian 8 and CentOS 7
  See #202

And, due to an API change, the driver's version has been bumped to `0.12.0`


Co-authored-by: Konstantin Germanov <kgermanov@axcient.com>