Implement support of kernel 6.0.x (#204)
It's required for our driver to support kernel 6.0.x. This patch covers this.

There are bio requests for special purposes, which are configured on the bio allocation by setting a specific bio
operations field. This field is represented by the enum req_op / enum req_ops. Historically, this enum was called
"req_op" (before v4.10), then it was renamed to "req_ops" (v.4.10+ ... v5.19). Previously, the elastio-snap driver
could handle both of these data types. But the logic worked that way that when in v6.0 that data type has been
renamed back to "req_op", this disabled the support of the "special" bio requests, which caused the memory
corruption.

 - Fixed the problem described above
 - Reverted patch to downgrade the kernel version on Fedora distros
 - Added compat for bdevname and REQ_OP_WRITE_ZEROES bio request
 - Rewritten alloc_disk() logic to make it more straightforward and fixed a refcnt warning

Closes #199