Add sytemd shutdown script for consistency of the rootfs (#206)
Tested on Ubuntu 20.04. The snapshots are failing to be reloaded due to 
a problem caused in initramfs script. It's logged on boot as info message:
```
[   34.92256]  EXT4-fs (sda5): INFO: recovery required on readonly filesystem 
```
Screenshot is here: https://user-images.githubusercontent.com/59954122/209341526-a60087e0-a7f7-4700-b2dd-695fd405d054.png

It happens because filesystem need to reapply journal and cannot do it in readonly mode.
We should remount rootfs to "ro" while shutdown at the very end to avoid this.