Create cow file precisely according to the user setting (#255)
Previously, the size of the COW file was the sum of the header size, mapping section, and the snapshot data itself.
This was added on purpose in one of the previous tasks and caused some issues with other tests. This approach
implicitly added some additional size to the fallocate specified by the user. It is now agreed that for system
administrators, it is more critical to precisely control the used disk space used, so we will bring the initial behavior
back. This means, that if the user specifies 10% of the disk for the COW file, this should be the precise size of the
COW file and no additional data should be added to it.

Changes:

* Brought back the previous logic of the COW file size calculation
* Added two CI tests that compare the expected and real COW file sizes

Closes #253