Implement debug utility for elastio-snap (#274)
This commit introduces the console debug utility for the elastio-snap driver. The design is the following:

 - driver sends Netlink events
 - the events are read by nl_debug (debugging utility)
 - events are proxied by using the UDP protocol
 - parsed and printed in the console if needed

The events are sent as multicast messages, so potential packet drops are possible. All messages have sequence numbers and nl_debug detects if any packet drops took place. If yes, it specifies the amount of missed packets. Despite the drops are possible, practically, they don't occur.

The utility supports different filtering that simplifies debugging:

 - bio in the sector range
 - bio of the specific type (read/write)
 - muting of different events by groups (COW file & bio events)
 - all messages are colored depending on the type to simplify reading (can be disabled)

The Netlink part is not compiled by default, use the following flag to include that part:

sudo make NETLINK_DEBUG=y

Closes #272